visionAddOn/firmware/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "pillow>=11.1.0",
    "pyqt6>=6.9.0",
    "scikit-image>=0.25.2",
]
